chore: moved SectionTitle into /section folder
